feat: make cuda graphs settable + move model opt
